Fix units on stake pool fees. (#221)
Paymetheus was mixing percentages ranged between 0-1 and 0-100 and
this was causing pool fees to be set too low.

+++ Paymetheus/ViewModels/PurchaseTicketsViewModel.cs
@@ -519,7 +519,7 @@ namespace Paymetheus.ViewModels
             {
                 purchaseResponse = await walletClient.PurchaseTicketsAsync(account, spendLimit,
                     requiredConfirms, votingAddress, _ticketsToPurchase, poolFeeAddress,
-                    (double)poolFees, expiryHeight, _splitFee, _ticketFee, passphrase);
+                    poolFees, expiryHeight, _splitFee, _ticketFee, passphrase);
             }
             catch (Grpc.Core.RpcException ex)
             {
